Interface - Revert User-Mode Extensions
This reverts the past three commits.

I'm not going to implement fixed values at user-level anymore. I think
that's dumb and too much work, and it'd make more sense to implement it
in the Set Records library. In addition, implementing stuff in the
library is more likely to have me not pulling my hair out over stuff.

So yeah. I'm going to start work on implementing it in the Set Records
library now. I'm keeping the reserved space because I still think it
could be nice to have a header of sorts, for checksums and things
perhaps.