Set Records - Variable and Fixed Set Segments
Implemented the backend stuff for fixed values in Set Records.
Basically, now there's a 'variable' segment, which is everything up to
and including the M-value, and then there's a 'fixed' segment which
follows that and has those predetermined values. All the logic has been
reworked to support this, and some variable names were changed to fit
the pattern better. I've tested this by creating a record with N = 5 and
M <= 31, so I could manually set a single fixed value of 32 and weed the
record. It seemed to work perfectly fine with the existing programs.

The next thing to do would be to get this to user-mode, what with
functions for getting information about the segments and what the fixed
values are, and getting this information into the Import/Export headers.