Set Records - Var./Fixed Segments at User Level
Added ability to control the length of the variable segment, as well as
the fixed values, through the Allocation function. Also these are stored
away in the file header for Import/Export. So now we have Records with
fixed values fully implemented. That was easy.

The programs were only changed insofar as to avoid breakage. They don't
yet support Records with fixed values, since they have to account for
the possibility that the M-range of a Record doesn't actually match up
with the M-values of sets. But really, it's only the Generation program
that needs to be reworked, as the Weed and Evaluate programs just read
from Query outputs.

I'm not going to go and implement something new in the Expand program to
deal with these fixed values. I think it's better to just imagine the
uppermost fixed value as the sole in-range M-value, and just filter out
sets that don't have a valid variable M-value. Eugh, I think maybe I
should make a new term for the variable segment's M-range, as it doesn't
directly correspond with set M-value anymore.