Generation - Basic Support for Fixed Values
The Generation program is now able to expand to records with fixed
values. It was already able to expand *from* them, and to enable this
the program just needed to understand that the 'M-value' given by the
library isn't necessarily the uppermost value in the set. So it's
working.

I think the last real change that needs to be made here is to give the
Create program the ability to make records with these fixed values.
Apart from that, maybe change the 'verbose' messages to give information
about fixed values. But yeah, almost done with this upgrade.