Generation - Fixed Values for Create Option
I forgot another minor change, which is that when the 'create' option is
used, it should use the same fixed values in the destination, because if
you remove them and keep the size it'd be enormous, and it's just
keeping consistent with how the option already works. The program won't
really be used like this, but just good to have it work consistently.