Set Records - Documentation, Variable Name
Added some documentation talking about the new change with Variable and
Fixed segments. M-value will still refer to the last value of the
Variable segment, since it just makes more sense to think of the Fixed
segment as being 'tacked on' to the variable set representations. Also
made a variable name consistent.