Updated Documentation
There's just one more significant upgrade I'd like to make before
this program is ready for mass computation:

- Exhaustive Test Upgrades
    - M-range of first reduction (as an equivalent to M-range of the
      source in a thorough expansion) replaces the range of the first
      computation thing
    - Mode parameter for enabling additive and multiplicative operations
      individually

Then I can start on designing the database generation process. It'll be
a script that automatically handles a whole bunch of Record files, doing
expansions and weedings to make the largest sets it can. It'll need to
like use compression on a lot of those boring Records, of course. It'll
need to keep track of what ranges of values have been fully cleared in
each one. It should have some way to automatically detect from progress
updates whether it should continue with a certain expansion or if it's
more time effective to just start weeding.

- Further Work
    - Exhaustive Test Optimizations
        - Mess about with individual compiler optimizations
        - Investigate a faster 'Weak Pass' with indeterminate result,
          maybe there's a super easy way to reduce the search space
        - Memoization or pattern recognition
        - Detect integer wrap-around (indeterminate)
    - Port to GPU or other accelerated compute device

Hoping I can at least get a start on that before the new semester
starts. Also hoping I'm able to find something interesting in all these
numbers.