Exhaustive Test - Initial Reduction Space
Implemented the M-range for restricting the space of the initial
reduction. This is a more general and useful version of what I
originally was doing by restricting the range of the 'first
computation.' This is to optimize the weeding process when there's been
a thorough expansion, since we know we don't need to reduce through one
of the sets we've already thoroughly expanded from. So we can do a
thorough expansion of a certain M-range, and that'd be identical to
weeding through that range.

Next I just need to implement this in the user-space, letting this be
controlled through the Weed program so it can be controlled by the mass
generation process.