Weed - Control of Initial Reduction Space
Implemented parameters in the Weed program for giving the M-range for
the initial reduction. Now the program can be invoked with a specific
range to weed by, and just covering all numbers with ranges in different
invocations gives a fully swept record.

I'm realizing that some of the original logic has become, well,
outdated, in a sense. Up until here, we've been able to skip doing
mutations on sets which are supersets of a nullifiable set, because we
know it'll have been mutated and expanded in anyways, because everything
is covered. But now, with doing smaller records that may only use a
specific M-range, we can't assume that anymore. In a source record, you
could have a superset of a set with a low M-value, whose mutations
aren't in-range of the record, so it'd be wrong not to perform mutations
on that set if we're calling it a 'thorough expansion.'

I'm thinking the only way to make this work and to let us have proper
thorough expansions is to disregard whether or not something's a
superset, and just do both supersets and mutations on everything. I
guess I could be overlooking something simpler, so I'll think over it
awhile. But what I'll probably do is just get rid of the superset
distinction in records entirely, which'd save space. I might also leave
it and just provide an option to be thorough, or to assume coverage.