Generation - Thorough Expansions, Documentation
Actually, what I'd been calling 'thorough expansions' does work with the
optimization still enabled. It's not that it has the same effect as
weeding that range, but given that all ranges are covered, the result is
the same. Basically, if we expand a swept record with that optimization,
there could be some of those 'previous mutations' out there that we need
to cover, but so long as we make the guarantee that in the end we'll
have every range expanded or weeded (to include those previous
mutations), we will still end up with a fully swept record.

But I think I need to switch up my terminology, since 'thorough
expansion' is what I'm now calling an expansion that doesn't use the
optimization. I added that feature because according to my reasoning
they should give the equivalent effect of a ranged weeding, as in
everything that can reduce to anything nullifiable in a certain range
will be marked after this. But in practice that isn't looking to be the
case. I'm getting slightly more marked sets with a ranged weeding than a
thorough expansion of a swept record. I'm expanding N = 4, 10-18 into an
N = 5, 1-20. Maybe it could be that the weeding program is overzealous.
I need to think about this to figure it out.