Generation - Redid Thorough Expansions
There's no way to absolutely get the same effect of a weeding through
thorough expansion. The reason is that there could be some sets that
reduce immediately to something with double-values. These are covered by
a weeding, but not necessarily by an expansion. But they are guaranteed
to be covered since really they're just supersets of length-3
nullifiable sets, similarly to the mutations of supersets thing. So
'thorough expansion' again means a fully executed expansion of a swept
record. Phew. Almost broke my terminology. Re-did the documentation to
discuss this in detail.