Interface - Issue with Large Records
So because the Record Query function returns a count of sets as an
ssize_t, it might be a really big number, too large for an int. This is
causing some issues with the error-check macros on really large records
over 2 GiB, since it loops around negative. Anyways, since large records
are pretty important for mass computation, I'm making this change.