Automatic Generation Script - Starting
I'm going to write a script to automatically generate a directory of
records. It should take in a source and destination, and a range of
M-values. It will automatically create appropriately sized records, then
perform thorough expansions into them, then weed them until they're
swept. Need this in order to do the database generation.