Database Script - Successor Range Chooser
Created a function that figures out the best successor range given the
previously generated one. Basically it will take in the set size, last
M-value, and fixed values, and spit out a new M-value range and new (or
the same) fixed values. It tries to make as large a record as possible
while keeping to a size restriction.

I need to try and document it better, like by making examples. Want to
show how the recursion helps with collapsing sequential fixed values, as
well as why the conditions are in this order.