Database Script - Filename Interpretation
Created a function to extract range information from a record filename.
It works. I don't think it's worth trying to rewrite or optimize. I
mean, it was pretty easy... it doesn't look amazing... but it's correct
and it gets the job done.

Next thing to do is probably work on something to scan a source
directory and expand into a target record. For now it can probably just
expand everything, but eventually I'll want to optimize it so it only
expands the necessary things -- if sets have more than one value
guaranteed to not be in the target, then it's not worth expanding, so we
can check records for this before expanding. Then after this it'd be
good to keep a log of what's been expanded into a particular record
file, or just the contiguous ranges.