Database Script - Automatic Expansion Function
Wrote the simple function to take a directory of record files and expand
them into a chosen destination record.

Change of plans. I think first I'll write a similar function to weed a
record, because in order to do the logging and stuff, which in the end
is just for ranged weeding, there needs to be a distinction between
swept and unswept records. And I think that distinction will be in the
filename. So yeah, first something to weed records so they're swept,
then can worry about logging, then can worry about skipping worthless
expansions.