Database Script - Tweaks, New Expansion Paradigm
I have a much more efficient plan for how to deal with expansions and
logging and stuff. An expansion will come from a directory of records
made by the script earlier. So these records will each just have an
associated index, and when expanding into a new record, it'll just keep
track of what index source it's currently at. Each directory will have
like a sorta TOC file that explains the M-range, so once a directory is
completely expanded in we can weed starting right after that range. It
could also be that 'log' for in-progress records, saving the current
source index. So yeah. That's a much nicer system.

So I implemented the record index as an additional prefix to the record
filename, with leading zeros for easy sorting. Then I also patched an
issue with the record creation function, where it just was incorrect for
fixed values (my dumb program uses variable and fixed segment sizes
individually and I keep forgetting). I implemented a basic Weed function
and re-implemented Expand similarly, so it just does one invocation of
the program. I'll write the loop in something that's aware of the
system. One last thing I did was fix the successor range calculator to
just halt when it gets to the given M-value limit.

Next step is to play around with putting stuff together. Maybe I'll do
the logging, maybe not. I'll probably just play around and see what I
can get working in an interactive shell. I've been able to work with
that pretty well so far, and everything seems to work correctly!