Database Script - NUMA Control, Mass Expansion
I created the system for mass expansion. It's a little bit messy here
and there, but overall it's easy to understand and work with and it gets
the job done. It's basically got a global counter for the index of the
next destination record, and there are multiple 'job' threads that are
working and will just pick up whatever the next one is and directory-
expand into it. Jobs have to stick to one NUMA node to avoid being super
inefficient, so I had to go back and implement that into the program
invocation routines.

A single Expansion is akin to a Dot Product (takes one source record and
expands into one destination record), a Directory Expansion is akin to a
Matrix-Vector Product (takes a source directory of records and expands
each into one destination record), meaning a Mass Expansion would be a
Matrix Product (takes a source directory and expands into a whole
destination directory).

I know... I've not really been following the plan. It keeps changing. I
think the next logical step is to make a similar process for sweeping a
directory, or maybe just weeding a certain range. Don't know. But what I
think I'll end up doing is reworking what I've just written so that the
thread stuff is sorta a wrapper that can be applied to a Mass Expansion
job or a Sweeping job. After that, I might want to look through and
clean stuff up, then implement directory logging, and then progress
logging on an invocation basis.