Database Script - Threading Wrapper, Working State
Reworked the threading stuff to be a wrapper around 'jobs.' The Mass
Processing routine will ensure a job is done across every record in a
destination directory. It does this by instantiating a bunch of worker
threads, split across NUMA nodes, each of which will look at the next
destination record and run a job on it. This job would in practice be a
directory expansion or a ranged weeding.

So now I have a fully functioning script that does mass expansions and
mass sweeping. Hooray! But there are still things I need to implement:

- File Compression
    - Implement option in program to not mark supersets specially, for
      even less information
- Directory Operation Logging
- Invocation Progress Logging/Resumption

Then there are some things I want to clean up:

- Ensure All Errors are Caught
- Reorder/Restructure Script, Column Restrictions