Database Script - Clean-up
I reordered a few things to make more sense. Also made individual
functions for the two mass processing modes, so they'll do the logging.
There has also been an issue with record ranges stopping immediately
before the max M, so I went and fixed that.

Here's the plan: the next thing to do is logging. I'll have the
individual mass processing functions handle this by outputting to the
directory log file. After this, I'll want to do the progress saves, so
I'll have the mass processing function run something to periodically
dump the indices of the next record and the current jobs, so if the
program were to crash it could resume by reinstating those jobs and the
next record index. Something I forgot to mention on the previous list
was to optimize the expansions by skipping ranges that just won't expand
into the destination range. After this, I'll implement file compression
in-place. It's an SSD, it can handle it.