Database Script - Operation Logging
Added bits to the mass processing functions to add a line to the log of
the destination directory. They're just meant to be human-readable. The
only thing the script itself will pick up on is the "INST"
(instantiation) line with N and the M-range, as well as if there's a
line saying "SWEPT," though this is only added manually (I don't want to
bother coding the thing to ascertain whether something's fully swept).

So I'm thinking too about having some more modes. I'm realizing there's
going to be a lot more human decisions and human interaction. I'm
thinking I'll separate supersets and mutations, since supersets can be
done more quickly and have the potential to get rid of a lot of stuff.
One might want to run supersets, then find that there are few enough
remaining sets to just weed right then and there. I'll also want to add
range to the weeding mode.