Database Script - Separate Supersets and Mutations
Created separate expansion modes for Supersets and Mutations. This is
helpful for reducing unnecessary work. Supersets are less costly, and so
can be done first, and if this leads to a low enough number of sets in
the destination (as it would near the minimum case), it could be that
mutations could be skipped.

Also re-specified the Weeding mode to have specific range.