Database Script - Invocation Resumption
Reworked the worker threads to read off current jobs from the array and
update them when finished. This way, job indices can be loaded in from
the progress file and worker threads will pick up jobs from there,
thereby resuming a previous invocation. I implemented this by checking
if the file exists and then prompting the user to optionally load it in.
I also changed up the progress save function from a thread routine to a
function that gets called each time there's an update.

Importantly, this resumption feature isn't a mode, but it activates
based on whether there is a progress file in the destination record,
something that's removed once an invocation fully completes. It's
important then to ensure the command and all parameters are the same
when using the feature.