Database Script - Record Compression for Mass Work
Implemented record (de)compression for mass processing. If a job is
scheduled on a record file who is compressed, it'll be automatically
decompressed for the job to run and then re-compressed afterward. Next
thing is to implement compression upon directory creation. After this,
the script is essentially ready for deployment.

Here are some other things to do just for quality-of-life, though:

- General clean-up and ensuring error handling is correct, column
  restrictions
- 'Inspect' mode for quickly looking at the unmarked set count in a
  directory... will figure out what works in deployment