Database Script - Bug Fix with Skipping
I deliberately initially coded the skipping logic this way (with the
skip toggles). I did this because when I was testing stuff out, I
noticed that just setting the do_x toggles would stop all activity on
the destination record, since it's a variable of a higher scope. Well, I
undid that a few commits ago in a "logic sweep." Better use this
version.