Database Script - Compression Levels, Configs
Record retrieval has been by index for jobs, so I made it that way for
getting expansion sources too. This enables other files to be stored in
the directory, like a configuration file I'm wanting to implement. This
would allow configuration of things like the number of jobs to run
concurrently, and how to compress files.

Speaking of compression, I redid it so that it's controlled by a global
constant. Having it be controlled by whether a file was compressed
before is problematic if the process gets interrupted. Also now it
allows for compression levels, as well as that one-bit marking mode I
implemented in the Generation program. I think every configuration I'll
want to make can be put in these global constant variables.