Database Script - Configuration Loading
Implemented config files in the form of JSON files. One can be loaded
from the directory the script is invoked from to supersede the defaults,
and additionally one can be loaded from the record directory itself to
supersede any of those. Basically, any configurations that are present
will be updated, and missing ones will remain unchanged.

I was also thinking about making it so that directory creation could use
existing directories and just delete everything except the config file,
so that a directory could be made with a specific config in it, before
instantiating it as a record directory. Maybe do it this way.