Database Script - Util Binary Dir, Conf Writing
Created another configuration that provides a location for the utility
binaries. Since these are getting a bit unwieldy and it's not easy to
memorize the things, I made it so that a config file with every
configurable field will be written to the invocation directory before
the record-specific one is loaded.