Database Script - Fixes
It wasn't a good idea to have a config get copied into a new directory.
That confused me. Let's just have it so that the user can copy it if
they please.

As for the 2-dimensional array for threads, well, I was experiencing an
issue where with multiple NUMA nodes configured, the progress file would
never get removed and strange things happened with inspections. It
looked like in that particular case the code would join the threads
before the worker function would exit. The real issue is that when that
list gets created, the sublists are actually all just references to the
same list, because Python is stupidly designed and shallow-copies the
hell out of everything. No normal person would say, "oh yeah, it'll copy
the pointer." So yeah, they'd end up being just the same thread objects
repeated, whichever ones were for the last node, and whenever these
finished was when the program would move on, assuming work was complete.