Database Script - Switched to Subprocess, Fix
There was a small issue where binary path wasn't being loaded in from
config files. Fixed this.

I've replaced all the calls to os.system() with calls to the Subprocess
module. I did this primarily because it's what's recommended. I'm
wanting to also implement progress tracking, and I think this might end
up working better with sending signals and stuff. But it might not make
any difference. Subprocess takes a list of arguments rather than a
command... and because of this I got a surprise when creating records
with more than one fixed value, since that program uses a
space-separated list. In a shell, you'd put quote marks around the list,
and they'll be ignored once we're in the actual program. So here we have
to make sure we don't put quote-marks. To keep the commands it prints
out executable, I created a function to just insert the quote-marks when
needed.