Database Script - Task Framework
Starting to rework the way that the script handles commands. So, jobs
are performed on individual target record files en masse. This involves
decompressing them (one at a time) once at the start and re-compressing
them afterwards. However, a Mass Processing routine involves doing all
the compressions and decompressions for just one task. And I'm finding
that the compressions can take quite a lot of CPU time. In a lot of
cases, it would be better to actually perform multiple tasks on each
record file, without decompressing and recompressing for each one. So it
could do multiple expansions, then a weeding, then an inspection, all in
one go, and all in one invocation too. So I've started the framework for
that, as adding tasks to a list to iterate on for each job (record
file). Next I will probably change up the way the commands themselves
are structured to allow for this.