Database Script - Multiple Task Commands
Reworked commands and command interpretation to allow for running
multiple tasks on a target at once. After specifying a destination
directory, the user can input a task character followed by task
arguments, and then another character for another task to follow it, and
so on. Next up would be progress tracking for jobs, if I still need it.