Database Script - Multiple Files, Separate Create
Separated the database script into a couple modules and split out
directory creation into its own script. Integrating record creation as a
task with all the others was proving problematic and would've involved
introducing a lot of inconsistency and things. The way I've engineered
this so far it makes more sense to simply deal with it separately.