Database Script - Inspection Tweak, Fix
Tweaked inspections to always output to a file in the directory being
inspected, with a particular filename format. This is just for better
reliability as filenames could collide if they're in the invocation
directory. Also fixed a logging error.

I'm pronouncing this version ready to be deployed again on the
workstation.