Database Script - Configuration Fixes
Upon running the program on the workstation for the first time, I
realized my splitting into multiple modules had broken the global vars
for configurations. So I fixed the imports and references to work
properly. Also, because they're unnecessary, I removed the job and
command function parameters for thread count.