Database Script - Repetitive M Inspections
Inspections would take a while when working with files with a larger
M-range as it has to run the Evaluate command for each one. So I've had
it now output to a tempfile in memory if there are more than two
M-values to do, and read in from it as opposed to running the command.
Hopefully this isn't terrible with memory usage, but I suppose it's only
really a concern for small N, as after some point we won't be finding
any unmarked sets in a record whose M-range is so low that it's large
enough to be problematic.