Exhaustive Test - No Wrap-Around, Equality Checks
In testing the N = 8 sets, I'm getting worried that there's potential
for wrap-around causing false negatives. If the test multiplies together
a lot of positive values, it could coincidentally line up with some
other values in the set and cause a zero to appear. Realistically
though, with the values I'm working with, it couldn't cause a false
positive. And it could be that even in principle that'd be impossible
due to the nature of wrap-around being well-defined... but I gotta do
some thinking on that. So what I've done here is simply disable using a
value that's wrapped around from multiplication. I guess technically now
the handful of results need to be tested by something else that can
handle large integers, though this doesn't matter for my purposes, as I
mentioned.

Also, I made what I think is an optimization on the checks for equality.
I'm doing them now right before any recursion takes place, potentially
saving lots of calls when there's already a computed zero. This means
the function in general doesn't take sets with repeated values. Also
enabled GCC optimization level 3, because I might as well. It should
speed things up. Once I've verified this new change works and is
correct, I'll deploy it and redo one of my expansions to check for
differences. I'd be surprised if there was anything different.

Since I've developed my system of nineteen formulations to check a
4-set, I think I might as well work on a replacement for this program.
I'll probably write it in Haskell as it's good for recursion and such,
and it has arbitrary-length integers. I also think I can manage to do
something similar for N = 5. I'd be interested to see how much runtime
these optimizations will save. Each of N = 2 and N = 3 saw a 5-6x speed-
up.