Evaluate - Multithreading
Implemented multithreading for the Evaluation command. When more than
one thread is used, the list it gives is no longer sorted. I haven't
been able to test this too much yet as it's just been on my laptop, but
I'm about to test this on the workstation. We'll see how much better it
is.