Evaluate, Interface -- High-Value Filters
Okay, so multithreading is fine, but multithreading while outputting
sets is not. It takes even longer. And yeah, I should've expected this.
So I've banned the printing of sets while multithreading, and I've
redone the Evaluate command to include the functionality from that
'Count' program.

Basically, it can now take in a list of M-values that it'll count up
occurrences of and output the numbers in order, just like Count. In
addition, you can specify some fixed values on top, since once we get
into larger records they'll have fixed values of their own, making
looking at the M-value useless. So we fix the top k values, and the
filters apply to a_{n-k}.

We can now go ahead and make changes in the script to remove the piping
of output and simply let this Evaluate command do the whole job, with
multithreading. Hopefully this should do the trick of speeding it up!