Database Script - Using Updated Eval
Now the script uses the filtering feature of the Evaluate command to do
the counting rather than piping to the extra program. And this does
indeed save a lot of time. I was trying a large N = 6 directory, which I
probably couldn't even reasonably do with 'grep's, took 28 minutes with
the extra program, and now 9 minutes, the majority of which was just
spent on file compression. So we're good.

But something weird is happening with multithreading. When filters are
set (specifically ones that will have sets counted towards it), and when
there are more of them, two-threaded performance seems to get worse than
single-threaded. I don't really know why. I changed it to atomic
operations and this seems faster, but still it doesn't scale. But
single-threaded is good enough for this.