Multiple Expand - Insertion Logic
Wrote the logic for iteratively and recursively inserting values into a
set to generate supersets. I think I can simplify it by making it take
an ending index as well. But this is kinda hard to understand as it is
and I was even struggling to understand what I'd done. This kind of
program is hard. Maybe I'll make something diagrammatic. From what I can
tell from a test I've done this implementation is correct.

Next I just want to try and polish up the logic, as well as finish the
user function for this library. After that I'll make a unit test program
and really test this.