Multiple Expand - Corrections, Test
I fixed some compile errors, as well as a stupid bug with input
validation. Also, there was a sneaky error in detecting if there is
already a working M-value, where it didn't account for the possibility
of two or more values being in the M-range. So now it only counts to
where values that can only be in the fixed range begin. From what I can
tell, this program gives correct output. However, I think it'd be even
better to have a separate program to automatically validate it,
something that can create the known correct supersets and make sure
they're all there and there's nothing else. Probably in Python.