Database Script - Special Supersets
Implemented the new General Multiple Superset utility in the Database
Script, as a special case of a regular superset expansion. There's no
change to the command being entered: the script will automatically
detect it's a size increase of more than one and use the multiple
superset feature. Surprisingly easy to modify this script! That's how I
know I've done well.

I tested this out by creating a directory of N = 5 sets for source
material, M = 100, immediately weeded. I expanded by supersets through
an intermediary N = 6 directory and into the target N = 7, and this took
almost two minutes (all in the latter expansion, the first was
negligible). Then I tried directly expanding from the source to the
target, and it took forty seconds and produced the same result.

This'll definitely be super helpful in re-generating and completing the
N = 8 database.