General Exhaustive Test - Recursive Reductions
Starting work on a new library that should provide a completely general
version of an exhaustive test. It can be a test for nullifiability
(having a bisectable (improper) subset), or perhaps even unifiability
(being bisectable or bisectasuccessive, without regard to subsets). It's
also gonna be generalized to testing any subset size, not just two, as
well as any base case size. We'll get to vary the parameters to see just
how much these techniques optimize the test!

The next thing to do is actually implement the tests for bisectability,
as well as a routine to check new subsets of a reduction. The cool thing
about this one is that sets are always going to be in ascending order,
so that'll let us optimize our tests. In particular if we're checking
subsets of length 2, then we only need to check at the left and right.
Lovely!