General Exhaustive Test - Tests, User Function
Implemented the logic for some of the tests, with some still to do. But
right now, this should have the same functionality of the original
nullifiability test. Also wrote the main function for the user.

Next I have to implement the bisectability test for N = 4 and then
routines to enumerate and check subsets of lengths 3 and 4. After that,
I want to generalize the test to anything besides bisectability,
implement other tests and introduce a parameter to select them.