General Exhaustive Test - Algorithm Correction
The big (couple-line) change here is that I realized there was a bit of
a mistake in the algorithm as written. The base case should also check
subsets not previously checked progressively. If we were progressively
checking N = 2 subsets, when we get to a N = 4 base, we would miss out
on any N = 3 subsets, which are important. We would in fact have N = 2
and fewer taken care of though, as that's the guarantee.

The other thing I've done is make a funny change that lets the subset-
checking routine check all subsets, in the case we don't have a
particular value we want to include. But I've done this very strangely
and roundaboutly, making a whole weird sorting routine for the first
case which isn't needed in the second at all. I'm gonna simplify the
code by making it all one case and going through N loops for an N-
subset, skipping whenever any prescribed value isn't included. That
might actually be faster than sorting each time. But I think this code
is somewhat cool so I'm just pushing it out here.