General Exhaustive Test - N = 4, Subset Checking
As mentioned previously, I changed the subset checking routine to have
an option to look at all subsets, not just ones containing a specific
value. This makes things easier for the preliminary check and the base
case test. I re-did it from last time so it's cleaner.

Also, I implemented the test for N = 4, all nineteen formulations. Let's
hope I got them right! I also put this in the subset checking routine.

So now I get to do a bunch of testing. I'm gonna try switching the Weed
program to this, play around with it, see how fast it is, and make sure
it gives correct results every time.