Weed - Removed Awful Slowdown
So this dumb feature has just been here slowing this whole thing down.
Mutexes. They are not instantaneous. The atomic operations do no better.
What's happening here is that for some low-N cases on the workstation,
when running this with more than one thread it would just do terribly,
all because the sets are too frequent, so there's always a wait to grab
that Mutex lock. This is why the multithreaded Evaluate util also
doesn't scale, but the whole point of that is to count anyways. So
whatever.

I came about this while testing out a version of the Weed command using
the General Exhaustive Test. It seemed correct, but slower, so maybe
should take a close look at the operations, maybe profile it. I'll push
the actual change here once I get to verify its correctness again.