Weed, Gen. Exh. Test - Switched Test, Correction
Switched the Weed program to using this new test library. I also changed
it to use the base case of N = 4 sets, and in the process discovered a
mistake in the subset-checking routine, preventing it from checking all
subsets, instead returning after the very first one. I also also changed
the filename.