Gen. Exh. Test - Avoid Reordered Parallel Ops
Introduced a new optimization I've been thinking about for a while now
and think I've figured out. It wasn't actually that hard. Basically,
there are situations where we would perform an operation on two values,
and then another on two different values. But we could do the same thing
in reverse order to get the same second-order reduction, and by default
our algorithm to enumerate reductions will do both methods. To solve
this, we must find a way to determine only one series of reduction
steps, and ensure the other doesn't happen. We do this by requiring a
second reduction step that is parallel, that is, only using values in
the original set, use a higher value than the first.

If we have the set {1,3,4,6,9} for example, we could reduce on the pair
1,6 to give {3,4,5,9} and then on 4,9 to get {3,5,5}. Since these were
parallel, each using values in the original set, we could have done it
in reverse order, first reducing on 4,9 to get {1,3,5,6} and then on 1,6
to get {3,5,5}. So let's try our system. Suppose we go down the path of
reducing on 1,6 first. We have the original values 3, 4, and 9
remaining, and since they all are above 1, we can operate between any of
them. Now suppose we go down the path of reducing on 4,9 first. We have
the original values 1, 3, and 6, and of those only 6 is higher than 4.
So there is no pair we can make according to our system, meaning the
parallel reduction on 1,6 won't be repeated. Hopefully that made sense.

So I tested this out for correctness by creating a record N = 5, M = 65,
and weeded it (which was incredibly fast). I then automatically
generated into the N = 6 successor record. Without the optimization,
weeding took 24 seconds on my desktop, and with the optimization, 18
seconds.

It'd be helpful when I write the block comment for this library to put a
list of all the optimizations I've made here. There have been many. And
some of them are particularly helpful in certain instances, which is
useful to make note of for the purpose of gathering data to show how
smart I am. For example, this optimization is probably more useful when
most sets to test are innullifiable, as it reduces the number of things
to exhaust.